refactor(ajaxHelper): remove separate ajax helper
-relates to #2045
