+++ spec/support/default.opts
@@ -1,5 +1,4 @@
 --require source-map-support/register
---require spec-js/helpers/ajax-helper.js
 --require spec-js/helpers/testScheduler-ui.js
 --ui spec-js/helpers/testScheduler-ui.js
